requirements.txt
# Core Data & Math
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Reinforcement Learning & ML
stable-baselines3[extra]>=2.1.0
gymnasium>=0.28.1
torch>=2.0.0

# Broker & Financial Data
alpaca-trade-api>=3.1.1
yfinance>=0.2.40
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0

# NLP & Sentiment
nltk>=3.8.1
textblob>=0.17.1

# Dashboard & Visualization
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0

# Infrastructure & Database
python-dotenv>=1.0.0
PyYAML>=6.0.1
redis>=5.0.0
psycopg2-binary>=2.9.9

# Utils
tqdm>=4.66.0
huggingface_hub>=0.19.0
